feat: create some components logic hooks
