+++ frontend/src/pages/CreateContactPage/index.tsx
@@ -1,30 +1,10 @@
-import { useRef } from "react";
-
-import ContactForm, { ContactFields, ContactFormRef } from "../../components/ContactForm";
+import ContactForm from "../../components/ContactForm";
 import PageHeader from "../../components/PageHeader";
-import toast from "../../components/Toast/toast";
-import ContactsService from "../../services/ContactsService";
+
+import { useCreateContact } from "./useCreateContact";
 
 function CreateContactPage() {
-  const contactFormRef = useRef<ContactFormRef>(null);
-
-  async function onSubmit(contact: ContactFields) {
-    try {
-      await ContactsService.create(contact);
-
-      contactFormRef.current?.resetForm();
-
-      toast({
-        title: "Contact registered successfully!",
-        variant: "success",
-      });
-    } catch {
-      toast({
-        title: "An error has occurred!",
-        variant: "danger",
-      });
-    }
-  }
+  const { contactFormRef, onSubmit } = useCreateContact();
 
   return (
     <>
