+++ frontend/src/pages/ListContactsPage/index.tsx
@@ -6,91 +6,38 @@ import {
   TrashSimple,
   WarningCircle,
 } from "@phosphor-icons/react";
-import { ChangeEvent, useCallback, useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 
 import { HStack, VStack } from "../../../styled-system/jsx";
 import { Button } from "../../components/Button";
-import { ContactFields } from "../../components/ContactForm";
 import Modal from "../../components/Modal";
 import PageLoading from "../../components/PageLoading";
 import SearchField from "../../components/SearchField";
-import toast from "../../components/Toast/toast";
-import ContactsService from "../../services/ContactsService";
 
 import * as S from "./styles";
-
-type ContactItem = ContactFields & { id: string; categoryName?: string };
+import { useListContacts } from "./useListContacts";
 
 function ListContactsPage() {
-  const [searchTerm, setSearchTerm] = useState<string>("");
-  const [orderByName, setOrderByName] = useState<"ASC" | "DESC">("ASC");
-  const [contacts, setContacts] = useState<ContactItem[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [hasError, setHasError] = useState(false);
-  const [isLoadingDelete, setIsLoadingDelete] = useState(false);
-  const [isDeleteModalVisible, setisDeleteModalVisible] = useState(false);
-  const [contactBeingDeleted, setContactBeingDeleted] = useState<ContactItem | null>(null);
+  const {
+    isLoading,
+    orderByName,
+    searchTerm,
+    contacts,
+    filteredContacts,
+    isLoadingDelete,
+    isDeleteModalVisible,
+    contactBeingDeleted,
+    hasError,
+    retry,
+    toggleOrderByName,
+    onConfirmDeleteContact,
+    onCancelDeleteContact,
+    onOpenDeleteContactModal,
+    onSearchFieldChange,
+  } = useListContacts();
 
-  const filteredContacts = contacts.filter((contact) =>
-    contact.name.toLowerCase().includes(searchTerm.toLowerCase()),
-  );
   // eslint-disable-next-line no-nested-ternary
   const headerAlignment = !hasError && contacts.length > 0 ? "space-between" : "center";
-
-  const fetchContacts = useCallback(async () => {
-    setIsLoading(true);
-    try {
-      const response = await ContactsService.list(orderByName);
-      setHasError(false);
-      setContacts(response);
-    } catch (e) {
-      setHasError(true);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [orderByName]);
-
-  useEffect(() => {
-    fetchContacts();
-  }, [fetchContacts]);
-
-  function toggleOrderByName() {
-    setOrderByName((prevState) => (prevState === "ASC" ? "DESC" : "ASC"));
-  }
-
-  function onSearchFieldChange(e: ChangeEvent<HTMLInputElement>) {
-    setSearchTerm(e.target.value);
-  }
-
-  function retry() {
-    fetchContacts();
-  }
-
-  async function onConfirmDeleteContact() {
-    setIsLoadingDelete(true);
-    try {
-      await ContactsService.delete(contactBeingDeleted!.id);
-
-      setContacts((prevState) =>
-        prevState.filter((contact) => contact.id !== contactBeingDeleted!.id),
-      );
-      setisDeleteModalVisible(false);
-      setContactBeingDeleted(null);
-
-      toast({
-        variant: "success",
-        title: "Contact deleted successfully",
-      });
-    } catch {
-      toast({
-        variant: "danger",
-        title: "An error occurred while deleting the contact",
-      });
-    } finally {
-      setIsLoadingDelete(false);
-    }
-  }
 
   return (
     <>
@@ -103,7 +50,7 @@ function ListContactsPage() {
         title={`Are you sure you want to remove the ${contactBeingDeleted?.name} contact?`}
         confirmLabel="Delete"
         onConfirm={onConfirmDeleteContact}
-        onCancel={() => setisDeleteModalVisible(false)}
+        onCancel={onCancelDeleteContact}
       >
         This action cannot be undone!
       </Modal>
@@ -186,10 +133,7 @@ function ListContactsPage() {
                           <button
                             type="button"
                             aria-label="Deletar Contato"
-                            onClick={() => {
-                              setisDeleteModalVisible(true);
-                              setContactBeingDeleted(contact);
-                            }}
+                            onClick={() => onOpenDeleteContactModal(contact)}
                           >
                             <TrashSimple />
                           </button>
